Fix two index out of bounds exceptions.
These trigger whenever trying to update a honmod.

In `MainForm.cs`:

```
The thread 0x21cc has exited with code 0 (0x0).
Exception thrown: 'System.IndexOutOfRangeException' in HoN_ModMan.exe
An unhandled exception of type 'System.IndexOutOfRangeException' occurred in HoN_ModMan.exe
Index was outside the bounds of the array.
```

In `ModUpdater.cs`:
```
Exception thrown: 'System.ArgumentException' in mscorlib.dll
System.ArgumentException: Offset and length were out of bounds for the array or count is greater than the number of elements from index to the end of the source collection.
   at System.IO.StreamReader.ReadBlock(Char[] buffer, Int32 index, Int32 count)
   at CS_ModMan.ModUpdater.UpdateThread() in D:\GitHub\Heroes-Of-Newerth-Mod-Manager\HON Mod Manager\ModUpdater.cs:line 158
The thread 0x4480 has exited with code 0 (0x0).
```