Fix up the logic for IsNewerVersion.
There were a few index out of bounds issues, which prevented the entire version string from being properly checked.

And some of the conditions were incorrect.

I cleaned this all up so that it properly checks every digit and returns `true` if there is a newer version. And `false` otherwise.